Implement lastsynced and unstablemsg commands
Implemented our lastsynced and unstablemsg commands. For .lastsynced, which can be run by any user, we will get our last synced from GetLastSynced. GetLastSynced will perform a lookup of the #Solus-Dev topic and assuming it is in the correct topic format, it will get the last set date via regexp and return that. If we fail at any point in our topic lookup, we'll just set it to "some point recently".

For unstablemsg, we follow our "Solus Development | Stable: Synced (Last on %s) | Unstable: %s" format, appending the information after .unstablemsg for unstable.

Need to implement some sort of "last checked" mechanism so we're not constantly querying for a new date, but rather only every hour or so, and force update when we do a .synced.